Moved distributed to non-default, implemented 97% accuracy model
